requirements.txt
# Pinned for reproducible builds and security audits. Run: pip-audit
fastapi>=0.115,<1
uvicorn>=0.32,<1
httpx>=0.27,<1
PyYAML>=6.0.1,<7
prometheus-client>=0.21,<1
python-json-logger>=2.0,<3
pydantic>=2,<3
pytest>=8,<9
